Fix tests

+++ tests/test_compiled.py
@@ -14,9 +14,9 @@ def make_ijk():
     kv = np.linspace(0.,nk-1., nk)
     i, j, k = np.meshgrid(iv, jv, kv, indexing='ij')
 
-    i = np.asfortranarray(np.expand_dims(i,-1))
-    j = np.asfortranarray(np.expand_dims(j,-1))
-    k = np.asfortranarray(np.expand_dims(k,-1))
+    i = np.asfortranarray(np.expand_dims(i,-1),dtype=np.float32)
+    j = np.asfortranarray(np.expand_dims(j,-1),dtype=np.float32)
+    k = np.asfortranarray(np.expand_dims(k,-1),dtype=np.float32)
 
     return i, j, k
 
@@ -31,9 +31,9 @@ def test_node_to_face():
     shape_iface = (ni, nj-1, nk-1, nv)
     shape_jface = (ni-1, nj, nk-1, nv)
     shape_kface = (ni-1, nj-1, nk, nv)
-    fi = np.empty(shape_iface,order='F')
-    fj = np.empty(shape_jface,order='F')
-    fk = np.empty(shape_kface,order='F')
+    fi = np.empty(shape_iface,order='F',dtype=np.float32)
+    fj = np.empty(shape_jface,order='F',dtype=np.float32)
+    fk = np.empty(shape_kface,order='F',dtype=np.float32)
 
     turbigen.compiled.node_to_face(fnode, fi, fj, fk)
 
@@ -62,20 +62,20 @@ def test_node_to_cell():
     xn = np.ones_like(i)
     ni, nj, nk, nv = i.shape
     shape_cell = (ni-1, nj-1, nk-1, nv)
-    xc = np.empty(shape_cell, order='F')
+    xc = np.empty(shape_cell, order='F', dtype=np.float32)
     turbigen.compiled.node_to_cell(xn, xc)
     assert np.allclose(xn[:-1,:-1,:-1,:], xc)
 
     # Error should be exactly half for linear variation in each dirn
-    ic = np.empty(shape_cell, order='F')
+    ic = np.empty(shape_cell, order='F', dtype=np.float32)
     turbigen.compiled.node_to_cell(i, ic)
     assert np.allclose(ic-i[:-1, :-1, :-1,:], 0.5)
 
-    jc = np.empty(shape_cell, order='F')
+    jc = np.empty(shape_cell, order='F', dtype=np.float32)
     turbigen.compiled.node_to_cell(j, jc)
     assert np.allclose(jc-j[:-1, :-1, :-1,:], 0.5)
 
-    kc = np.empty(shape_cell, order='F')
+    kc = np.empty(shape_cell, order='F', dtype=np.float32)
     turbigen.compiled.node_to_cell(k, kc)
     assert np.allclose(kc-k[:-1, :-1, :-1,:], 0.5)
 
@@ -89,24 +89,24 @@ def test_cell_to_node():
     xc = np.ones_like(i)
     ni, nj, nk, nv = xc.shape
     shape_node = (ni+1, nj+1, nk+1, nv)
-    xn = np.empty(shape_node, order='F')
+    xn = np.empty(shape_node, order='F', dtype=np.float32)
     turbigen.compiled.cell_to_node(xc, xn)
     assert np.allclose(xc, 1.)
 
     # Check linear variation in each direction
-    inode = np.empty(shape_node, order='F')
+    inode = np.empty(shape_node, order='F', dtype=np.float32)
     turbigen.compiled.cell_to_node(i, inode)
     assert np.allclose(inode[0,:-1,:-1], i[0,:,:])
     assert np.allclose(inode[-1,:-1,:-1], i[-1,:,:])
     assert np.allclose(inode[1:-1,:-1,:-1]-i[:-1,:,:],0.5)
 
-    jnode = np.empty(shape_node, order='F')
+    jnode = np.empty(shape_node, order='F', dtype=np.float32)
     turbigen.compiled.cell_to_node(j, jnode)
     assert np.allclose(jnode[:-1,0,:-1], j[:,0,:])
     assert np.allclose(jnode[:-1,-1,:-1], j[:,-1,:])
     assert np.allclose(jnode[:-1,1:-1,:-1]-j[:,:-1,:],0.5)
 
-    knode = np.empty(shape_node, order='F')
+    knode = np.empty(shape_node, order='F', dtype=np.float32)
     turbigen.compiled.cell_to_node(k, knode)
     assert np.allclose(knode[:-1,:-1,0], k[:,:,0])
     assert np.allclose(knode[:-1,:-1,-1], k[:,:,-1])
@@ -117,7 +117,7 @@ def test_smooth_zero():
     # Zero smoothing factor should change nothing
     shape = (2,4,5,6)
     X = np.random.random_sample(shape)
-    Xs = np.asfortranarray(X.copy())
+    Xs = np.asfortranarray(X.copy()).astype(np.float32)
     turbigen.compiled.smooth(Xs, sf2=0., sf4=0.)
     assert np.allclose(X, Xs)
 
@@ -126,7 +126,7 @@ def test_smooth_const():
     # A constant value should stay constant after smoothing
     for sf2 in (0.1, 0.2):
         for sf4 in (0.1, 0.2):
-            X = np.ones((10,15,20,1),order='F')
+            X = np.ones((10,15,20,1),order='F', dtype=np.float32)
             turbigen.compiled.smooth(X, sf2=sf2, sf4=sf4)
             assert np.allclose(X, 1.)
 
@@ -179,7 +179,7 @@ def test_smooth_converge():
     sf4 = 0.3
     derr = np.inf
     for istep in range(10000):
-        Xnew = np.asfortranarray(X.copy())
+        Xnew = np.asfortranarray(X.copy()).astype(np.float32)
         turbigen.compiled.smooth(Xnew, sf2, sf4)
         derr = X.ptp() - Xnew.ptp()
         X = Xnew
